% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift-package.R
\docType{package}
\name{redshift-package}
\alias{redshift}
\alias{redshift-package}
\title{redshift: Utilites for AWS Redshift}
\description{
Utilites for working with AWS Redshift.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mur@craftsy.com}

}
\keyword{internal}
