% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_accept_reserved_node_exchange}
\alias{redshift_accept_reserved_node_exchange}
\title{Exchanges a DC1 Reserved Node for a DC2 Reserved Node with no changes to
the configuration (term, payment type, or number of nodes) and no
additional costs}
\usage{
redshift_accept_reserved_node_exchange(ReservedNodeId,
  TargetReservedNodeOfferingId)
}
\arguments{
\item{ReservedNodeId}{[required] A string representing the node identifier of the DC1 Reserved Node to be
exchanged.}

\item{TargetReservedNodeOfferingId}{[required] The unique identifier of the DC2 Reserved Node offering to be used for
the exchange. You can obtain the value for the parameter by calling
GetReservedNodeExchangeOfferings}
}
\description{
Exchanges a DC1 Reserved Node for a DC2 Reserved Node with no changes to
the configuration (term, payment type, or number of nodes) and no
additional costs.
}
\section{Request syntax}{
\preformatted{redshift$accept_reserved_node_exchange(
  ReservedNodeId = "string",
  TargetReservedNodeOfferingId = "string"
)
}
}

\keyword{internal}
