% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_operations.R
\name{redshift_cancel_resize}
\alias{redshift_cancel_resize}
\title{Cancels a resize operation}
\usage{
redshift_cancel_resize(ClusterIdentifier)
}
\arguments{
\item{ClusterIdentifier}{[required] The unique identifier for the cluster that you want to cancel a resize
operation for.}
}
\description{
Cancels a resize operation.
}
\section{Request syntax}{
\preformatted{redshift$cancel_resize(
  ClusterIdentifier = "string"
)
}
}

\keyword{internal}
