% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.new.R
\name{s3copy}
\alias{s3copy}
\title{Load data from R data frame to Redshift via a COPY (from S3) command.}
\usage{
s3copy(df, redshift_conn, schema_name, table_name, iam_role_arn, s3_bucket,
  s3_key, s3_delete_on_success = TRUE, time_tz = NULL,
  time_tz_offset = TRUE, time_digits = 6, gz = TRUE)
}
